pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openevals"
version = "1.0.0"
description = "An open-source evaluation framework for large language models"
readme = "README.md"
authors = [{ name = "Hailey Cheng" }]
license = { text = "MIT License" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"

dependencies = [
    "torch>=2.4.0,<3.0.0",
    "transformers>=4.45.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",
    "scipy>=1.14.0",
    "PyYAML>=6.0.2",
    "pydantic>=2.9.0",
    "tqdm>=4.66.0",
    "pandas>=2.2.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.9.0",
    "psutil>=6.0.0",
    "nvidia-ml-py3>=7.352.0",
    "huggingface-hub>=0.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=4.0.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "flake8>=7.1.0",
    "mypy>=1.13.0",
    "pip-tools>=7.4.0",
    "numpy>=2.0.0",
    "ruff>=0.8.0"
]

[project.scripts]
run-benchmark = "openevals.scripts.run_benchmark:main"
download-data = "openevals.scripts.download_data:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["openevals*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["openevals"]
skip = ["venv", ".venv", "node_modules"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = ["venv", ".venv", "node_modules", ".git"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "gpu: marks tests as requiring GPU",
    "integration: marks tests as integration tests",
]
